Changed open_rate from Int to Double
This variable is causing the following error while being an Int:

java.lang.IllegalArgumentException: com.google.gson.JsonSyntaxException: java.lang.NumberFormatException: Expected an int but was 40.20618556701 at line 1 column 6339 path $.lists[1].stats.open_rate

+++ src/main/java/com/ecwid/maleorang/method/v3_0/lists/StatsInfo.kt
@@ -59,7 +59,7 @@ class StatsInfo : MailchimpObject() {
 
     @JvmField
     @Field
-    var open_rate: Int? = null
+    var open_rate: Double? = null
 
     @JvmField
     @Field
